pkm_gardener/requirements.txt
google-generativeai
PyYAML
PyMuPDF
Pillow
python-dotenv
reportlab
